pyproject.toml
[tool.poetry]
name = "whisper-bot"
version = "0.1.0"
description = ""
authors = ["saycoron <saycoron@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ruff = "^0.1.2"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"
openai = "^0.28.1"
pydub = "^0.25.1"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
